+++ src/game/tests.rs
@@ -552,6 +552,7 @@ mod counted_bodies;
 mod counter_conditions;
 mod countering_and_mana;
 mod creature_bond;
+mod cyclone;
 mod damage_dealt_by;
 mod damage_history;
 mod damage_limits;